put SAS in one file
