--- modules/e2ee/SAS.ts
@@ -1,66 +1,0 @@
-export const emojiMapping = [
-    ["🐶", "dog"],
-    ["🐱", "cat"],
-    ["🦁", "lion"],
-    ["🐎", "horse"],
-    ["🦄", "unicorn"],
-    ["🐷", "pig"],
-    ["🐘", "elephant"],
-    ["🐰", "rabbit"],
-    ["🐼", "panda"],
-    ["🐓", "rooster"],
-    ["🐧", "penguin"],
-    ["🐢", "turtle"],
-    ["🐟", "fish"],
-    ["🐙", "octopus"],
-    ["🦋", "butterfly"],
-    ["🌷", "flower"],
-    ["🌳", "tree"],
-    ["🌵", "cactus"],
-    ["🍄", "mushroom"],
-    ["🌏", "globe"],
-    ["🌙", "moon"],
-    ["☁️", "cloud"],
-    ["🔥", "fire"],
-    ["🍌", "banana"],
-    ["🍎", "apple"],
-    ["🍓", "strawberry"],
-    ["🌽", "corn"],
-    ["🍕", "pizza"],
-    ["🎂", "cake"],
-    ["❤️", "heart"],
-    ["🙂", "smiley"],
-    ["🤖", "robot"],
-    ["🎩", "hat"],
-    ["👓", "glasses"],
-    ["🔧", "spanner"],
-    ["🎅", "santa"],
-    ["👍", "thumbs up"],
-    ["☂️", "umbrella"],
-    ["⌛", "hourglass"],
-    ["⏰", "clock"],
-    ["🎁", "gift"],
-    ["💡", "light bulb"],
-    ["📕", "book"],
-    ["✏️", "pencil"],
-    ["📎", "paperclip"],
-    ["✂️", "scissors"],
-    ["🔒", "lock"],
-    ["🔑", "key"],
-    ["🔨", "hammer"],
-    ["☎️", "telephone"],
-    ["🏁", "flag"],
-    ["🚂", "train"],
-    ["🚲", "bicycle"],
-    ["✈️", "aeroplane"],
-    ["🚀", "rocket"],
-    ["🏆", "trophy"],
-    ["⚽", "ball"],
-    ["🎸", "guitar"],
-    ["🎺", "trumpet"],
-    ["🔔", "bell"],
-    ["⚓️", "anchor"],
-    ["🎧", "headphones"],
-    ["📁", "folder"],
-    ["📌", "pin"],
-];
