+++ modules/e2ee/crypto-workers.ts
@@ -1,4 +1,3 @@
-import { emojiMapping } from "./SAS";
 import { hash } from 'internxt-crypto';
 
 
@@ -36,3 +35,70 @@ export function logWarning(message: string) {
 export function logError(message: string) {
     console.error(`E2E: ${message}`);
 }
+
+ export const emojiMapping = [
+    ["🐶", "dog"],
+    ["🐱", "cat"],
+    ["🦁", "lion"],
+    ["🐎", "horse"],
+    ["🦄", "unicorn"],
+    ["🐷", "pig"],
+    ["🐘", "elephant"],
+    ["🐰", "rabbit"],
+    ["🐼", "panda"],
+    ["🐓", "rooster"],
+    ["🐧", "penguin"],
+    ["🐢", "turtle"],
+    ["🐟", "fish"],
+    ["🐙", "octopus"],
+    ["🦋", "butterfly"],
+    ["🌷", "flower"],
+    ["🌳", "tree"],
+    ["🌵", "cactus"],
+    ["🍄", "mushroom"],
+    ["🌏", "globe"],
+    ["🌙", "moon"],
+    ["☁️", "cloud"],
+    ["🔥", "fire"],
+    ["🍌", "banana"],
+    ["🍎", "apple"],
+    ["🍓", "strawberry"],
+    ["🌽", "corn"],
+    ["🍕", "pizza"],
+    ["🎂", "cake"],
+    ["❤️", "heart"],
+    ["🙂", "smiley"],
+    ["🤖", "robot"],
+    ["🎩", "hat"],
+    ["👓", "glasses"],
+    ["🔧", "spanner"],
+    ["🎅", "santa"],
+    ["👍", "thumbs up"],
+    ["☂️", "umbrella"],
+    ["⌛", "hourglass"],
+    ["⏰", "clock"],
+    ["🎁", "gift"],
+    ["💡", "light bulb"],
+    ["📕", "book"],
+    ["✏️", "pencil"],
+    ["📎", "paperclip"],
+    ["✂️", "scissors"],
+    ["🔒", "lock"],
+    ["🔑", "key"],
+    ["🔨", "hammer"],
+    ["☎️", "telephone"],
+    ["🏁", "flag"],
+    ["🚂", "train"],
+    ["🚲", "bicycle"],
+    ["✈️", "aeroplane"],
+    ["🚀", "rocket"],
+    ["🏆", "trophy"],
+    ["⚽", "ball"],
+    ["🎸", "guitar"],
+    ["🎺", "trumpet"],
+    ["🔔", "bell"],
+    ["⚓️", "anchor"],
+    ["🎧", "headphones"],
+    ["📁", "folder"],
+    ["📌", "pin"],
+];
